# openapi.yaml
openapi: 3.1.0
info:
  title: News AI Agent API
  description: "\n    News AI Agent REST API\n    \n    A production-ready API for\
    \ managing RSS feeds, articles, and triggering AI-powered workflows.\n    \n \
    \   ## Features\n    \n    - **Feed Management**: CRUD operations for RSS feed\
    \ sources\n    - **Article Management**: Query and manage news articles\n    -\
    \ **Workflows**: Trigger n8n workflows for scraping, summarization, and email\
    \ digests\n    - **System Monitoring**: Health checks and statistics\n    \n \
    \   ## Authentication\n    \n    Most endpoints require an API key in the `X-API-Key`\
    \ header.\n    The `/api/health` endpoint is public and does not require authentication.\n\
    \    \n    ## Documentation\n    \n    - **Swagger UI**: `/docs` - Interactive\
    \ API documentation\n    - **ReDoc**: `/redoc` - Alternative documentation format\n\
    \    - **OpenAPI JSON**: `/openapi.json` - Machine-readable API specification\n\
    \    "
  version: 1.0.0
paths:
...
  /api/articles:
    get:
      tags:
      - Article Management
      summary: List Articles
      description: List articles with optional filters.
      operationId: list_articles_api_articles_get
      parameters:
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by source name (partial match)
          title: Source
        description: Filter by source name (partial match)
      - name: processed
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter by processing status
          title: Processed
        description: Filter by processing status
      - name: sent
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter by sent status
          title: Sent
        description: Filter by sent status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of articles to return
          default: 50
          title: Limit
        description: Maximum number of articles to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of articles to skip
          default: 0
          title: Offset
        description: Number of articles to skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
                title: Response List Articles Api Articles Get
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/articles/{article_id}:
    get:
      tags:
      - Article Management
      summary: Get Article
      description: Get a specific article by ID.
      operationId: get_article_api_articles__article_id__get
      parameters:
      - name: article_id
        in: path
        required: true
        schema:
          type: string
          title: Article Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Article not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Article Management
      summary: Delete Article Endpoint
      description: Delete an article.
      operationId: delete_article_endpoint_api_articles__article_id__delete
      parameters:
      - name: article_id
        in: path
        required: true
        schema:
          type: string
          title: Article Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Delete Article Endpoint Api Articles  Article Id  Delete
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Article not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/recipients:
    post:
      tags:
      - Recipient Management
      summary: Add Recipient
      description: Add a new email recipient for news digests. Can be an individual
        email or a distribution list.
      operationId: create_recipient_api_recipients_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipientCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '400':
          description: Recipient with this email already exists
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Recipient Management
      summary: List Recipients
      description: List all digest recipients. Use enabled_only=true to filter to
        active recipients.
      operationId: list_recipients_api_recipients_get
      parameters:
      - name: enabled_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Enabled Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipient'
                title: Response List Recipients Api Recipients Get
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/recipients/{recipient_id}:
    get:
      tags:
      - Recipient Management
      summary: Get Recipient
      description: Get a specific recipient by ID.
      operationId: get_recipient_api_recipients__recipient_id__get
      parameters:
      - name: recipient_id
        in: path
        required: true
        schema:
          type: string
          title: Recipient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Recipient not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Recipient Management
      summary: Update Recipient
      description: Update an existing recipient's email, name, or enabled status.
      operationId: update_recipient_api_recipients__recipient_id__put
      parameters:
      - name: recipient_id
        in: path
        required: true
        schema:
          type: string
          title: Recipient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipientUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Recipient not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Recipient Management
      summary: Delete Recipient
      description: Permanently remove a recipient from the system.
      operationId: delete_recipient_api_recipients__recipient_id__delete
      parameters:
      - name: recipient_id
        in: path
        required: true
        schema:
          type: string
          title: Recipient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Recipient not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/recipients/{recipient_id}/toggle:
    patch:
      tags:
      - Recipient Management
      summary: Toggle Recipient
      description: Toggle a recipient's enabled status (enable/disable without deletion).
      operationId: toggle_recipient_api_recipients__recipient_id__toggle_patch
      parameters:
      - name: recipient_id
        in: path
        required: true
        schema:
          type: string
          title: Recipient Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Recipient not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Article:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        url:
          type: string
          title: Url
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
        source:
          type: string
          title: Source
        source_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Type
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
        fetched_at:
          type: string
          format: date-time
          title: Fetched At
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        sentiment:
          anyOf:
          - type: string
          - type: 'null'
          title: Sentiment
        sentiment_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Sentiment Score
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
        priority_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority Reason
        topics:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Topics
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
        companies:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Companies
        persons:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Persons
        regions:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Regions
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        relevance_match:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Relevance Match
        processed:
          type: boolean
          title: Processed
        sent:
          type: boolean
          title: Sent
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - title
      - url
      - content
      - source
      - source_type
      - published_at
      - fetched_at
      - summary
      - sentiment
      - sentiment_score
      - priority
      - priority_reason
      - topics
      - keywords
      - companies
      - persons
      - regions
      - language
      - category
      - relevance_match
      - processed
      - sent
      - created_at
      - updated_at
      title: Article
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Recipient:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        enabled:
          type: boolean
          title: Enabled
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - name
      - enabled
      - created_at
      - updated_at
      title: Recipient
      description: Digest recipient response model.
    RecipientCreate:
      properties:
        email:
          type: string
          maxLength: 255
          minLength: 5
          pattern: ^[^@]+@[^@]+\.[^@]+$
          title: Email
        name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Name
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - email
      title: RecipientCreate
      description: Create a new digest recipient.
    RecipientUpdate:
      properties:
        email:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 5
            pattern: ^[^@]+@[^@]+\.[^@]+$
          - type: 'null'
          title: Email
        name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Name
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      type: object
      title: RecipientUpdate
      description: Update an existing recipient.
...
    SuccessResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - message
      title: SuccessResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Get from your .env file.
servers:
- url: http://localhost:3000
  description: Local development
- url: https://api.example.com
  description: Production (update this)
security:
- APIKeyHeader: []